Add online check
Until now, qb was not checking the network connection before starting
any borg processes. A prominent error is when the system wakes up from
sleep and immediately triggers a timed `qb run` command, where some or
all of the contained steps might fail.
Now, a ping is sent (and retried if failing) before the actual process
starts. However, this is off by default and mus be explicitly
configured.